requirements.txt
blenderproc==2.3.0
certifi==2021.10.8
chamferdist==1.0.0
charset-normalizer==2.0.12
click==8.1.3
cupy-cuda113==10.5.0
cycler==0.11.0
docker-pycreds==0.4.0
easydict==1.9
efficientnet-pytorch==0.7.1
fastrlock==0.8
fonttools==4.33.3
future==0.18.2
future-fstrings==1.2.0
fvcore==0.1.5.post20220512
gitdb==4.0.9
GitPython==3.1.27
h5py==3.7.0
idna==3.3
imageio==2.21.2
iopath==0.1.10
joblib==1.1.0
kiwisolver==1.4.2
kornia==0.6.5
lpips-pytorch===latest
matplotlib==3.5.2
munch==2.5.0
numpy==1.22.4
opencv-contrib-python==4.6.0.66
OpenEXR==1.3.9
packaging==21.3
pathtools==0.1.2
Pillow==9.1.0
piqa==1.2.2
portalocker==2.6.0
pretrainedmodels==0.7.4
progressbar==2.5
promise==2.3
protobuf==3.20.1
psutil==5.9.1
PyEXR==0.3.10
pyparsing==3.0.8
python-dateutil==2.8.2
pytorch3d==0.7.1
PyYAML==6.0
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.0
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tabulate==0.9.0
termcolor==2.1.0
threadpoolctl==3.1.0
timm==0.5.4
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
typing==3.7.4.3
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.12.17
yacs==0.1.8
